refactor(chat-api): rename features/chat → features/streaming
The old `chat/` feature was replaced by `conversations/`; the directory now
holds only the SSE/run-event streaming plumbing, so the name was misleading.

- features/chat/ → features/streaming/
- chat.streaming.ts → sse-sender.ts, chat.events.ts → events.ts,
  chat.logger.ts → logger.ts (run-event-sink.ts / run-events-to-sse.ts keep
  their names)
- ChatLogger → RequestLogger (it's a per-request logger used by conversations
  and sandbox-orchestration, not chat-specific)

Pure rename + import updates; no behavior change. AGENTS.md updated.
